InterconnectingConverter

PowerSystems.InterconnectingConverter — Type

mutable struct InterconnectingConverter <: StaticInjection
    name::String
    available::Bool
    bus::ACBus
    dc_bus::DCBus
    active_power::Float64
    rating::Float64
    active_power_limits::MinMax
    base_power::Float64
    loss_function::Union{LinearCurve, QuadraticCurve}
    services::Vector{Service}
    dynamic_injector::Union{Nothing, DynamicInjection}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

Interconnecting Power Converter (IPC) for transforming power from an ACBus to a DCBus

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bus::ACBus: Bus on the AC side of this converter

dc_bus::DCBus: Bus on the DC side of this converter

active_power::Float64: Active power (MW) on the DC side, validation range: active_power_limits

rating::Float64: Maximum output power rating of the converter (MVA), validation range: (0, nothing)

active_power_limits::MinMax: Minimum and maximum stable active power levels (MW)

base_power::Float64: Base power of the converter in MVA, validation range: (0, nothing)

loss_function::Union{LinearCurve, QuadraticCurve}: (default: LinearCurve(0.0)) Linear or quadratic loss function with respect to the converter current

services::Vector{Service}: (default: Device[]) Services that this device contributes to

dynamic_injector::Union{Nothing, DynamicInjection}: (default: nothing) corresponding dynamic injection device

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::InterconnectingConverter) -> String

Get InterconnectingConverter name.

PowerSystems.get_active_power — Method

get_active_power(value::InterconnectingConverter) -> Any

Get InterconnectingConverter active_power.

PowerSystems.get_active_power_limits — Method

get_active_power_limits(
    value::InterconnectingConverter
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get InterconnectingConverter active_power_limits.

PowerSystems.get_available — Method

get_available(value::InterconnectingConverter) -> Bool

Get InterconnectingConverter available.

PowerSystems.get_base_power — Method

get_base_power(value::InterconnectingConverter) -> Float64

Get InterconnectingConverter base_power.

PowerSystems.get_bus — Method

get_bus(value::InterconnectingConverter) -> ACBus

Get InterconnectingConverter bus.

PowerSystems.get_dc_bus — Method

get_dc_bus(value::InterconnectingConverter) -> DCBus

Get InterconnectingConverter dc_bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::InterconnectingConverter
) -> Union{Nothing, DynamicInjection}

Get InterconnectingConverter dynamic_injector.

PowerSystems.get_ext — Method

get_ext(
    value::InterconnectingConverter
) -> Dict{String, Any}

Get InterconnectingConverter ext.

PowerSystems.get_loss_function — Method

get_loss_function(
    value::InterconnectingConverter
) -> Union{LinearCurve, QuadraticCurve}

Get InterconnectingConverter loss_function.

PowerSystems.get_rating — Method

get_rating(value::InterconnectingConverter) -> Any

Get InterconnectingConverter rating.

PowerSystems.get_services — Method

get_services(
    value::InterconnectingConverter
) -> Vector{Service}

Get InterconnectingConverter services.

PowerSystems.set_active_power! — Method

set_active_power!(
    value::InterconnectingConverter,
    val
) -> Any

Set InterconnectingConverter active_power.

PowerSystems.set_active_power_limits! — Method

set_active_power_limits!(
    value::InterconnectingConverter,
    val
) -> Any

Set InterconnectingConverter active_power_limits.

PowerSystems.set_available! — Method

set_available!(value::InterconnectingConverter, val) -> Any

Set InterconnectingConverter available.

PowerSystems.set_base_power! — Method

set_base_power!(value::InterconnectingConverter, val) -> Any

Set InterconnectingConverter base_power.

PowerSystems.set_bus! — Method

set_bus!(value::InterconnectingConverter, val) -> Any

Set InterconnectingConverter bus.

PowerSystems.set_dc_bus! — Method

set_dc_bus!(value::InterconnectingConverter, val) -> Any

Set InterconnectingConverter dc_bus.

PowerSystems.set_ext! — Method

set_ext!(value::InterconnectingConverter, val) -> Any

Set InterconnectingConverter ext.

PowerSystems.set_loss_function! — Method

set_loss_function!(
    value::InterconnectingConverter,
    val
) -> Any

Set InterconnectingConverter loss_function.

PowerSystems.set_rating! — Method

set_rating!(value::InterconnectingConverter, val) -> Any

Set InterconnectingConverter rating.

PowerSystems.set_services! — Method

set_services!(value::InterconnectingConverter, val) -> Any

Set InterconnectingConverter services.