TModelHVDCLine

PowerSystems.TModelHVDCLine — Type

mutable struct TModelHVDCLine <: DCBranch
    name::String
    available::Bool
    active_power_flow::Float64
    arc::Arc
    r::Float64
    l::Float64
    c::Float64
    active_power_limits_from::MinMax
    active_power_limits_to::MinMax
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A High Voltage DC transmission line for modeling DC transmission networks.

This line must be connected to a DCBus on each end. It uses a T-Model of the line impedance. This is suitable for operational simulations with a multi-terminal DC network

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow on the line (MW)

arc::Arc: An Arc defining this line from a bus to another bus

r::Float64: Total series Resistance in p.u. (SYSTEM_BASE), split equally on both sides of the shunt capacitance

l::Float64: Total series Inductance in p.u. (SYSTEM_BASE), split equally on both sides of the shunt capacitance

c::Float64: Shunt capacitance in p.u. (SYSTEM_BASE)

active_power_limits_from::MinMax: Minimum and maximum active power flows to the FROM node (MW)

active_power_limits_to::MinMax: Minimum and maximum active power flows to the TO node (MW)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::TModelHVDCLine) -> String

Get TModelHVDCLine name.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::TModelHVDCLine) -> Any

Get TModelHVDCLine active_power_flow.

PowerSystems.get_active_power_limits_from — Method

get_active_power_limits_from(
    value::TModelHVDCLine
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get TModelHVDCLine active_power_limits_from.

PowerSystems.get_active_power_limits_to — Method

get_active_power_limits_to(
    value::TModelHVDCLine
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get TModelHVDCLine active_power_limits_to.

PowerSystems.get_arc — Method

get_arc(value::TModelHVDCLine) -> Arc

Get TModelHVDCLine arc.

PowerSystems.get_available — Method

get_available(value::TModelHVDCLine) -> Bool

Get TModelHVDCLine available.

PowerSystems.get_c — Method

get_c(value::TModelHVDCLine) -> Float64

Get TModelHVDCLine c.

PowerSystems.get_ext — Method

get_ext(value::TModelHVDCLine) -> Dict{String, Any}

Get TModelHVDCLine ext.

PowerSystems.get_l — Method

get_l(value::TModelHVDCLine) -> Float64

Get TModelHVDCLine l.

PowerSystems.get_r — Method

get_r(value::TModelHVDCLine) -> Float64

Get TModelHVDCLine r.

PowerSystems.get_services — Method

get_services(value::TModelHVDCLine) -> Vector{Service}

Get TModelHVDCLine services.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(value::TModelHVDCLine, val) -> Any

Set TModelHVDCLine active_power_flow.

PowerSystems.set_active_power_limits_from! — Method

set_active_power_limits_from!(
    value::TModelHVDCLine,
    val
) -> Any

Set TModelHVDCLine active_power_limits_from.

PowerSystems.set_active_power_limits_to! — Method

set_active_power_limits_to!(
    value::TModelHVDCLine,
    val
) -> Any

Set TModelHVDCLine active_power_limits_to.

PowerSystems.set_arc! — Method

set_arc!(value::TModelHVDCLine, val) -> Any

Set TModelHVDCLine arc.

PowerSystems.set_available! — Method

set_available!(value::TModelHVDCLine, val) -> Any

Set TModelHVDCLine available.

PowerSystems.set_c! — Method

set_c!(value::TModelHVDCLine, val) -> Any

Set TModelHVDCLine c.

PowerSystems.set_ext! — Method

set_ext!(value::TModelHVDCLine, val) -> Any

Set TModelHVDCLine ext.

PowerSystems.set_l! — Method

set_l!(value::TModelHVDCLine, val) -> Any

Set TModelHVDCLine l.

PowerSystems.set_r! — Method

set_r!(value::TModelHVDCLine, val) -> Any

Set TModelHVDCLine r.

PowerSystems.set_services! — Method

set_services!(value::TModelHVDCLine, val) -> Any

Set TModelHVDCLine services.