Transformer2W

PowerSystems.Transformer2W — Type

mutable struct Transformer2W <: ACBranch
    name::String
    available::Bool
    active_power_flow::Float64
    reactive_power_flow::Float64
    arc::Arc
    r::Float64
    x::Float64
    primary_shunt::Float64
    rating::Union{Nothing, Float64}
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A basic 2-winding transformer.

The model uses an equivalent circuit assuming the impedance is on the High Voltage Side of the transformer. The model allocates the iron losses and magnetizing susceptance to the primary side

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow through the transformer (MW)

reactive_power_flow::Float64: Initial condition of reactive power flow through the transformer (MVAR)

arc::Arc: An Arc defining this transformer from a bus to another bus

r::Float64: Resistance in pu (SYSTEM_BASE), validation range: (-2, 4)

x::Float64: Reactance in pu (SYSTEM_BASE), validation range: (-2, 4)

primary_shunt::Float64: Shunt reactance in pu (SYSTEM_BASE), validation range: (0, 2)

rating::Union{Nothing, Float64}: Thermal rating (MVA). Flow through the transformer must be between -rating and rating. When defining a transformer before it is attached to a System, rating must be in per-unit divided by the base power of the System it will be attached to, validation range: (0, nothing)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::Transformer2W) -> String

Get Transformer2W name.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::Transformer2W) -> Any

Get Transformer2W active_power_flow.

PowerSystems.get_arc — Method

get_arc(value::Transformer2W) -> Arc

Get Transformer2W arc.

PowerSystems.get_available — Method

get_available(value::Transformer2W) -> Bool

Get Transformer2W available.

PowerSystems.get_ext — Method

get_ext(value::Transformer2W) -> Dict{String, Any}

Get Transformer2W ext.

PowerSystems.get_primary_shunt — Method

get_primary_shunt(value::Transformer2W) -> Float64

Get Transformer2W primary_shunt.

PowerSystems.get_r — Method

get_r(value::Transformer2W) -> Float64

Get Transformer2W r.

PowerSystems.get_rating — Method

get_rating(value::Transformer2W) -> Any

Get Transformer2W rating.

PowerSystems.get_reactive_power_flow — Method

get_reactive_power_flow(value::Transformer2W) -> Any

Get Transformer2W reactive_power_flow.

PowerSystems.get_services — Method

get_services(value::Transformer2W) -> Vector{Service}

Get Transformer2W services.

PowerSystems.get_x — Method

get_x(value::Transformer2W) -> Float64

Get Transformer2W x.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(value::Transformer2W, val) -> Any

Set Transformer2W active_power_flow.

PowerSystems.set_arc! — Method

set_arc!(value::Transformer2W, val) -> Any

Set Transformer2W arc.

PowerSystems.set_available! — Method

set_available!(value::Transformer2W, val) -> Any

Set Transformer2W available.

PowerSystems.set_ext! — Method

set_ext!(value::Transformer2W, val) -> Any

Set Transformer2W ext.

PowerSystems.set_primary_shunt! — Method

set_primary_shunt!(value::Transformer2W, val) -> Any

Set Transformer2W primary_shunt.

PowerSystems.set_r! — Method

set_r!(value::Transformer2W, val) -> Any

Set Transformer2W r.

PowerSystems.set_rating! — Method

set_rating!(value::Transformer2W, val) -> Any

Set Transformer2W rating.

PowerSystems.set_reactive_power_flow! — Method

set_reactive_power_flow!(value::Transformer2W, val) -> Any

Set Transformer2W reactive_power_flow.

PowerSystems.set_services! — Method

set_services!(value::Transformer2W, val) -> Any

Set Transformer2W services.

PowerSystems.set_x! — Method

set_x!(value::Transformer2W, val) -> Any

Set Transformer2W x.