TwoTerminalHVDCLine

PowerSystems.TwoTerminalHVDCLine — Type

mutable struct TwoTerminalHVDCLine <: ACBranch
    name::String
    available::Bool
    active_power_flow::Float64
    arc::Arc
    active_power_limits_from::MinMax
    active_power_limits_to::MinMax
    reactive_power_limits_from::MinMax
    reactive_power_limits_to::MinMax
    loss::NamedTuple{(:l0, :l1), Tuple{Float64, Float64}}
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A High Voltage DC line, which must be connected to an ACBus on each end.

This model is appropriate for operational simulations with a linearized DC power flow approximation with losses proportional to the power flow. For modeling a DC network, see TModelHVDCLine

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow on the line (MW)

arc::Arc: An Arc defining this line from a bus to another bus

active_power_limits_from::MinMax: Minimum and maximum active power flows to the FROM node (MW)

active_power_limits_to::MinMax: Minimum and maximum active power flows to the TO node (MW)

reactive_power_limits_from::MinMax: Minimum and maximum reactive power limits to the FROM node (MVAR)

reactive_power_limits_to::MinMax: Minimum and maximum reactive power limits to the TO node (MVAR)

loss::NamedTuple{(:l0, :l1), Tuple{Float64, Float64}}: Linear loss model coefficients, where l0 = constant loss (MW) and l1 = linearly proportional loss rate (MW of loss per MW of flow)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::TwoTerminalHVDCLine) -> String

Get TwoTerminalHVDCLine name.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::TwoTerminalHVDCLine) -> Any

Get TwoTerminalHVDCLine active_power_flow.

PowerSystems.get_active_power_limits_from — Method

get_active_power_limits_from(
    value::TwoTerminalHVDCLine
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get TwoTerminalHVDCLine active_power_limits_from.

PowerSystems.get_active_power_limits_to — Method

get_active_power_limits_to(
    value::TwoTerminalHVDCLine
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get TwoTerminalHVDCLine active_power_limits_to.

PowerSystems.get_arc — Method

get_arc(value::TwoTerminalHVDCLine) -> Arc

Get TwoTerminalHVDCLine arc.

PowerSystems.get_available — Method

get_available(value::TwoTerminalHVDCLine) -> Bool

Get TwoTerminalHVDCLine available.

PowerSystems.get_ext — Method

get_ext(value::TwoTerminalHVDCLine) -> Dict{String, Any}

Get TwoTerminalHVDCLine ext.

PowerSystems.get_loss — Method

get_loss(
    value::TwoTerminalHVDCLine
) -> @NamedTuple{l0::Float64, l1::Float64}

Get TwoTerminalHVDCLine loss.

PowerSystems.get_reactive_power_limits_from — Method

get_reactive_power_limits_from(
    value::TwoTerminalHVDCLine
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get TwoTerminalHVDCLine reactive_power_limits_from.

PowerSystems.get_reactive_power_limits_to — Method

get_reactive_power_limits_to(
    value::TwoTerminalHVDCLine
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get TwoTerminalHVDCLine reactive_power_limits_to.

PowerSystems.get_services — Method

get_services(value::TwoTerminalHVDCLine) -> Vector{Service}

Get TwoTerminalHVDCLine services.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(
    value::TwoTerminalHVDCLine,
    val
) -> Any

Set TwoTerminalHVDCLine active_power_flow.

PowerSystems.set_active_power_limits_from! — Method

set_active_power_limits_from!(
    value::TwoTerminalHVDCLine,
    val
) -> Any

Set TwoTerminalHVDCLine active_power_limits_from.

PowerSystems.set_active_power_limits_to! — Method

set_active_power_limits_to!(
    value::TwoTerminalHVDCLine,
    val
) -> Any

Set TwoTerminalHVDCLine active_power_limits_to.

PowerSystems.set_arc! — Method

set_arc!(value::TwoTerminalHVDCLine, val) -> Any

Set TwoTerminalHVDCLine arc.

PowerSystems.set_available! — Method

set_available!(value::TwoTerminalHVDCLine, val) -> Any

Set TwoTerminalHVDCLine available.

PowerSystems.set_ext! — Method

set_ext!(value::TwoTerminalHVDCLine, val) -> Any

Set TwoTerminalHVDCLine ext.

PowerSystems.set_loss! — Method

set_loss!(value::TwoTerminalHVDCLine, val) -> Any

Set TwoTerminalHVDCLine loss.

PowerSystems.set_reactive_power_limits_from! — Method

set_reactive_power_limits_from!(
    value::TwoTerminalHVDCLine,
    val
) -> Any

Set TwoTerminalHVDCLine reactive_power_limits_from.

PowerSystems.set_reactive_power_limits_to! — Method

set_reactive_power_limits_to!(
    value::TwoTerminalHVDCLine,
    val
) -> Any

Set TwoTerminalHVDCLine reactive_power_limits_to.

PowerSystems.set_services! — Method

set_services!(value::TwoTerminalHVDCLine, val) -> Any

Set TwoTerminalHVDCLine services.