ValueCurves

PowerSystems.LinearCurve — Type

LinearCurve(proportional_term::Float64)
LinearCurve(proportional_term::Float64, constant_term::Float64)

A linear input-output curve, representing a constant marginal rate. May have zero no-load cost (i.e., constant average rate) or not.

Arguments

proportional_term::Float64: marginal rate

constant_term::Float64: optional, cost at zero production, defaults to 0.0

PowerSystems.PiecewiseAverageCurve — Type

PiecewiseAverageCurve(initial_input::Union{Float64, Nothing}, x_coords::Vector{Float64}, slopes::Vector{Float64})

A piecewise linear curve specified by average rates between production points. May have nonzero initial value.

Arguments

initial_input::Union{Float64, Nothing}: cost at minimum production point first(x_coords) (NOT at zero production), defines the start of the curve

x_coords::Vector{Float64}: vector of n production points

slopes::Vector{Float64}: vector of n-1 average rates/slopes of the curve segments between the points

PowerSystems.PiecewiseIncrementalCurve — Type

PiecewiseIncrementalCurve(initial_input::Union{Float64, Nothing}, x_coords::Vector{Float64}, slopes::Vector{Float64})
PiecewiseIncrementalCurve(input_at_zero::Union{Nothing, Float64}, initial_input::Union{Float64, Nothing}, x_coords::Vector{Float64}, slopes::Vector{Float64})

A piecewise linear curve specified by marginal rates (slopes) between production points. May have nonzero initial value.

Arguments

input_at_zero::Union{Nothing, Float64}: (optional, defaults to nothing) cost at zero production, does NOT represent a part of the curve

initial_input::Union{Float64, Nothing}: cost at minimum production point first(x_coords) (NOT at zero production), defines the start of the curve

x_coords::Vector{Float64}: vector of n production points

slopes::Vector{Float64}: vector of n-1 marginal rates/slopes of the curve segments between the points

PowerSystems.PiecewisePointCurve — Type

PiecewisePointCurve(points::Vector{Tuple{Float64, Float64}})

A piecewise linear curve specified by cost values at production points.

Arguments

points::Vector{Tuple{Float64, Float64}} or similar: vector of (production, cost) pairs

PowerSystems.QuadraticCurve — Type

QuadraticCurve(quadratic_term::Float64, proportional_term::Float64, constant_term::Float64)

A quadratic input-output curve, may have nonzero no-load cost.

Arguments

quadratic_term::Float64: quadratic term of the curve

proportional_term::Float64: proportional term of the curve

constant_term::Float64: constant term of the curve

InfrastructureSystems.get_constant_term — Method

get_constant_term(vc::LinearCurve) -> Float64

Get the constant term (i.e., intercept) of the LinearCurve

InfrastructureSystems.get_constant_term — Method

get_constant_term(vc::QuadraticCurve) -> Float64

Get the constant term of the QuadraticCurve

InfrastructureSystems.get_points — Method

get_points(
    vc::PiecewisePointCurve
) -> Vector{@NamedTuple{x::Float64, y::Float64}}

Get the points that define the PiecewisePointCurve

InfrastructureSystems.get_proportional_term — Method

get_proportional_term(vc::LinearCurve) -> Float64

Get the proportional term (i.e., slope) of the LinearCurve

InfrastructureSystems.get_proportional_term — Method

get_proportional_term(vc::QuadraticCurve) -> Float64

Get the proportional (i.e., linear) term of the QuadraticCurve

InfrastructureSystems.get_quadratic_term — Method

get_quadratic_term(vc::QuadraticCurve) -> Float64

Get the quadratic term of the QuadraticCurve

InfrastructureSystems.get_slopes — Method

get_slopes(vc::PiecewiseIncrementalCurve) -> Vector{Float64}

Fetch the slopes that define the PiecewiseIncrementalCurve

InfrastructureSystems.get_slopes — Method

get_slopes(vc::PiecewisePointCurve) -> Vector{Float64}

Calculate the slopes of the line segments defined by the PiecewisePointCurve

InfrastructureSystems.get_x_coords — Method

get_x_coords(vc::PiecewiseAverageCurve) -> Vector{Float64}

Get the x-coordinates that define the PiecewiseAverageCurve

InfrastructureSystems.get_x_coords — Method

get_x_coords(
    vc::PiecewiseIncrementalCurve
) -> Vector{Float64}

Get the x-coordinates that define the PiecewiseIncrementalCurve

InfrastructureSystems.get_x_coords — Method

get_x_coords(vc::PiecewisePointCurve) -> Vector{Float64}

Get the x-coordinates of the points that define the PiecewisePointCurve

InfrastructureSystems.get_y_coords — Method

get_y_coords(vc::PiecewisePointCurve) -> Vector{Float64}

Get the y-coordinates of the points that define the PiecewisePointCurve

PowerSystems.get_average_rates — Method

get_average_rates(
    vc::PiecewiseAverageCurve
) -> Vector{Float64}

Get the average rates that define the PiecewiseAverageCurve